separate spanner impls, templetize, remove aatop
