+++ vowpalwabbit/core/src/reductions/cb/details/large_action/vanilla_rand_svd_impl.cc
@@ -200,7 +200,7 @@ void vanilla_rand_svd_impl::run(const multi_ex& examples, const std::vector<floa
   }
 }
 
-vanilla_rand_svd_impl::vanilla_rand_svd_impl(VW::workspace* all, uint64_t d, uint64_t seed)
+vanilla_rand_svd_impl::vanilla_rand_svd_impl(VW::workspace* all, uint64_t d, uint64_t seed, size_t)
     : _all(all), _d(d), _seed(seed)
 {
 }
